requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
psycopg[binary]>=3.2
python-dotenv>=1.0
celery>=5.4
redis>=5.0
django-cors-headers>=4.3,<5